[PredNet] Change variable back to float
If not changed to float, it may occur problems regarding stacking different dtypes.

+++ PredNet/train.py
@@ -54,7 +54,7 @@ def train(model, optim, schedule, lossp, dataloader, device, logger,
         # run through all batches in the dataloader
         for batch_id, data in enumerate(dataloader[0]):
             # get sequence and target
-            x = data.to(device).permute(1,0,2,3,4)
+            x = data.to(device).float().permute(1,0,2,3,4)
             if norm or binar:
                 x = normalize(x)
             if binar:
@@ -81,7 +81,7 @@ def train(model, optim, schedule, lossp, dataloader, device, logger,
 
             # log scalar values
             logger.plot_loss('error', loss, it)
-            
+
             if batch_id >= iteration and iteration > 0:
                 break
